+++ src/Program.cs
@@ -75,6 +75,7 @@ internal class Program
             new AzureMcp.Areas.Sql.SqlSetup(),
             new AzureMcp.Areas.Storage.StorageSetup(),
             new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
+            new AzureMcp.Areas.AzureTerraformBestPractices.AzureTerraformBestPracticesSetup(),
         ];
     }